+++ packages/whisky-csl/src/tx_parser/certificates.rs
@@ -1,0 +1,9 @@
+use whisky_common::{Certificate, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn certificates(&mut self) -> Result<Vec<Certificate>, WError> {
+        Ok(vec![])
+    }
+}
