+++ packages/whisky-csl/src/tx_parser/change_address.rs
@@ -1,0 +1,9 @@
+use whisky_common::WError;
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn change_address(&mut self) -> Result<String, WError> {
+        Ok("".to_string())
+    }
+}
