+++ packages/whisky-csl/src/tx_parser/change_datum.rs
@@ -1,0 +1,9 @@
+use whisky_common::{Datum, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn change_datum(&mut self) -> Result<Option<Datum>, WError> {
+        Ok(None)
+    }
+}
