+++ packages/whisky-csl/src/tx_parser/collaterals.rs
@@ -1,0 +1,9 @@
+use whisky_common::{PubKeyTxIn, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn collaterals(&mut self) -> Result<Vec<PubKeyTxIn>, WError> {
+        Ok(vec![])
+    }
+}
