+++ packages/whisky-csl/src/tx_parser/inputs.rs
@@ -1,0 +1,9 @@
+use whisky_common::{TxIn, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn inputs(&mut self) -> Result<Vec<TxIn>, WError> {
+        Ok(vec![])
+    }
+}
