+++ packages/whisky-csl/src/tx_parser/metadata.rs
@@ -1,0 +1,9 @@
+use whisky_common::{Metadata, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn metadata(&mut self) -> Result<Vec<Metadata>, WError> {
+        Ok(vec![])
+    }
+}
