+++ packages/whisky-csl/src/tx_parser/mints.rs
@@ -1,0 +1,9 @@
+use whisky_common::{MintItem, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn mints(&mut self) -> Result<Vec<MintItem>, WError> {
+        Ok(vec![])
+    }
+}
