+++ packages/whisky-csl/src/tx_parser/reference_inputs.rs
@@ -1,0 +1,9 @@
+use whisky_common::{RefTxIn, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn reference_inputs(&mut self) -> Result<Vec<RefTxIn>, WError> {
+        Ok(vec![])
+    }
+}
