+++ packages/whisky-csl/src/tx_parser/votes.rs
@@ -1,0 +1,9 @@
+use whisky_common::{Vote, WError};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn votes(&mut self) -> Result<Vec<Vote>, WError> {
+        Ok(vec![])
+    }
+}
