+++ packages/whisky-csl/src/tx_parser/withdrawals.rs
@@ -1,0 +1,9 @@
+use whisky_common::{WError, Withdrawal};
+
+use super::TxParser;
+
+impl TxParser {
+    pub fn withdrawals(&mut self) -> Result<Vec<Withdrawal>, WError> {
+        Ok(vec![])
+    }
+}
